requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
setuptools>=58.0.0
django-cors-headers>=3.13.0,<4.0
bitcoinlib